Theme-mode selector, rules kept between mode, begining of cssRules
